backend/file_sharing/requirements.txt
django
django-cors-headers
psycopg2
djangorestframework
pip install djangorestframework-simplejwt
